Convert only by article number
This provides a more flexible and realistic parsing of the original URL
path. A regular expression is used to match the language and article
number while clearly marking the optional components of the path. In
particular, since the slug suffix is optional, the conversion is now
done entirely from the Zendesk article number to wiki.js title slug.